Fix MapBox and MapLibre zooming
MapBox has a feature for detecting mouse wheel vs trackpads that we
sometimes misclassify our injected events as mouse wheels, causing
annoying slow zooming. Split the events for those to have smaller
deltas to get MapBox to always treat them as trackpads.

Also change inject_main to be injected when map is loaded so that it
can be loaded on delayed-initialized iframes (and also save injecting
if no maps are detected)